+++ runtime/vitreus/src/lib.rs
@@ -125,7 +125,8 @@ use sp_consensus_beefy::{
 };
 use sp_runtime::transaction_validity::InvalidTransaction;
 use vitreus_runtime_common::{
-    ExposureMultiplier, NativeEnergyExchange, QuotePriceEnergyForNative, QuotePriceNativeForEnergy,
+    ExposureMultiplier, NativeEnergyExchange, QuotePrice, QuotePriceEnergyForNative,
+    QuotePriceNativeForEnergy,
 };
 use xcm::{
     latest::prelude::AssetId as XcmAssetId, VersionedAssetId, VersionedAssets, VersionedLocation,
@@ -254,7 +255,7 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
     spec_name: create_runtime_str!("vitreus-power-plant"),
     impl_name: create_runtime_str!("vitreus-power-plant"),
     authoring_version: 1,
-    spec_version: 210,
+    spec_version: 211,
     impl_version: 0,
     apis: RUNTIME_API_VERSIONS,
     transaction_version: 4,
@@ -488,12 +489,17 @@ impl pallet_assets::Config for Runtime {
     type MetadataDepositPerByte = MetadataDepositPerByte;
     type ApprovalDeposit = ApprovalDeposit;
     type StringLimit = AssetsStringLimit;
-    type Freezer = ();
+    type Freezer = AssetsFreezer;
     type Extra = ();
     type CallbackHandle = ();
     type WeightInfo = pallet_assets::weights::SubstrateWeight<Runtime>;
     #[cfg(feature = "runtime-benchmarks")]
     type BenchmarkHelper = ();
+}
+
+impl pallet_assets_freezer::Config for Runtime {
+    type RuntimeFreezeReason = RuntimeFreezeReason;
+    type RuntimeEvent = RuntimeEvent;
 }
 
 impl pallet_reputation::Config for Runtime {
@@ -931,7 +937,7 @@ impl pallet_assets::Config<PoolAssetsInstance> for Runtime {
     type BenchmarkHelper = ();
 }
 
-type NativeOrAssetId = frame_support::traits::fungible::NativeOrWithId<AssetId>;
+pub type NativeOrAssetId = frame_support::traits::fungible::NativeOrWithId<AssetId>;
 
 pub type NativeAndAssets = frame_support::traits::fungible::UnionOf<
     Balances,
@@ -1809,6 +1815,7 @@ construct_runtime!(
         TransactionPayment: pallet_transaction_payment = 7,
         Sudo: pallet_sudo = 8,
         PoolAssets: pallet_assets::<Instance1> = 9,
+        AssetsFreezer: pallet_assets_freezer = 10,
 
         EVM: pallet_evm = 15,
         EVMChainId: pallet_evm_chain_id = 16,
@@ -2693,13 +2700,33 @@ impl_runtime_apis! {
         }
     }
 
-    impl energy_broker_runtime_api::EnergyBrokerApi<Block, Balance> for Runtime {
+    impl energy_broker_runtime_api::EnergyBrokerApi<Block, AccountId, NativeOrAssetId, Balance> for Runtime {
         fn estimate_energy_from_native(amount: Balance) -> Option<Balance> {
             <EnergyBrokerExchange as QuotePriceNativeForEnergy>::quote_price_exact_tokens_for_tokens(amount, true)
         }
 
         fn estimate_native_from_energy(amount: Balance) -> Option<Balance> {
             <EnergyBrokerExchange as QuotePriceEnergyForNative>::quote_price_exact_tokens_for_tokens(amount, true)
+        }
+
+        fn quote_price_exact_tokens_for_tokens(
+            _who: Option<AccountId>,
+            asset1: NativeOrAssetId,
+            asset2: NativeOrAssetId,
+            amount: Balance,
+            include_fee: bool,
+        ) -> Option<Balance> {
+            EnergyBroker::quote_price_exact_tokens_for_tokens(asset1, asset2, amount, include_fee)
+        }
+
+        fn quote_price_tokens_for_exact_tokens(
+            _who: Option<AccountId>,
+            asset1: NativeOrAssetId,
+            asset2: NativeOrAssetId,
+            amount: Balance,
+            include_fee: bool,
+        ) -> Option<Balance> {
+            EnergyBroker::quote_price_tokens_for_exact_tokens(asset1, asset2, amount, include_fee)
         }
 
         fn energy_exchange_rate() -> Option<FixedU128> {