+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog for the Vitreus runtime.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [211] - 2025-02-15
+
+### Added
+
+- Add `pallet-assets-freezer`
+- Add `quote_price_exact_tokens_for_tokens` and `quote_price_tokens_for_exact_tokens` to EnergyBroker runtime API
+
 ## [210] - 2025-02-11
 
 ### Added